requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-Script==2.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyMySQL==1.0.2
python-decouple==3.6
Werkzeug==2.2.3
